requirements.txt
Flask~=3.1.0
flasgger~=0.9.7.1
PyYAML~=6.0.2